# Conflicts:
#	project/SPTarkov.Core/Helpers/HttpHelper.cs
#	project/SPTarkov.Core/SPT_Data/Launcher/Locales/Chinese (Simplified).json
#	project/SPTarkov.Core/SPT_Data/Launcher/Locales/Chinese (Traditional).json
#	project/SPTarkov.Core/SPT_Data/Launcher/Locales/English.json
#	project/SPTarkov.Core/SPT_Data/Launcher/Locales/French.json
#	project/SPTarkov.Core/SPT_Data/Launcher/Locales/German.json
#	project/SPTarkov.Core/SPT_Data/Launcher/Locales/Italian.json
#	project/SPTarkov.Core/SPT_Data/Launcher/Locales/Japanese.json
#	project/SPTarkov.Core/SPT_Data/Launcher/Locales/Korean.json
#	project/SPTarkov.Core/SPT_Data/Launcher/Locales/Polish.json
#	project/SPTarkov.Core/SPT_Data/Launcher/Locales/Russian.json
#	project/SPTarkov.Core/SPT_Data/Launcher/Locales/Spanish.json
#	project/SPTarkov.Core/SPT_Data/Launcher/Locales/Turkish.json
#	project/SPTarkov.Launcher/Components/Forge/DetailedModComponent.razor
#	project/SPTarkov.Launcher/Dialogs/Forge/ForgeUpdateDialog.razor
#	project/SPTarkov.Launcher/Pages/Forge.razor

+++ project/SPTarkov.Core/Mods/UpdateTask.cs
@@ -13,7 +13,7 @@ public class UpdateTask : IModTask
     public required string GUID { get; set; }
     public required string Link { get; set; }
     public float Progress { get; set; }
-    public float TotalToDownload { get; set; }
+    public long TotalToDownload { get; set; }
     public required CancellationTokenSource CancellationTokenSource { get; set; }
     public bool Complete { get; set; }
     public Exception? Error { get; set; }
